backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
anthropic>=0.42.0
tenacity>=9.0.0
structlog>=24.0.0
python-dotenv>=1.0.0
